refactor: работа с озвучкой и голосом
