+++ src/assistantSdk/voice/player/trackStream.ts
@@ -1,65 +1,4 @@
-/** Создает структуру для хранения загружаемых и воспроизводимых частей трека */
-const createChunkQueue = <T extends AudioBufferSourceNode>() => {
-    const buffer: Array<T> = []; // очередь на воспроизведение
-    const chunks: Array<T> = []; // очередь воспроизведения
-
-    let duration = 0; // продолжительность очереди на воспроизведение
-    let loaded = false; // флаг завершения загрузки
-
-    /** Добавить чанк в очередь на воспроизведение */
-    const push = (chunk: T) => {
-        buffer.push(chunk);
-        duration += chunk.buffer?.duration || 0;
-    };
-
-    /** Добавить чанк в очередь воспроизведения */
-    const toPlay = (chunk: T) => {
-        chunks.push(chunk);
-    };
-
-    /** Удалить чанк из очереди воспроизведения */
-    const remove = (chunk: T) => {
-        chunks.splice(chunks.indexOf(chunk), 1);
-    };
-
-    /** Получить очередь на воспроизведение */
-    const popAll = (): Array<T> => {
-        duration = 0;
-        return buffer.splice(0, buffer.length);
-    };
-
-    /** Проставляем признак окончания загрузки трека */
-    const allLoaded = () => {
-        loaded = true;
-    };
-
-    return {
-        get bufferLen() {
-            return buffer.length;
-        },
-        get chunks() {
-            return chunks;
-        },
-        allLoaded,
-        toPlay,
-        remove,
-        push,
-        popAll,
-        get length() {
-            return chunks.length;
-        },
-        get duration() {
-            return duration;
-        },
-        get ended() {
-            // считаем трек законченным, когда все загружено и воспроизведено
-            return loaded && chunks.length === 0 && buffer.length === 0;
-        },
-        get loaded() {
-            return loaded;
-        },
-    };
-};
+import { createChunkQueue } from './chunkQueue';
 
 type BytesArraysSizes = {
     incomingMessageVoiceDataLength: number;
@@ -77,7 +16,7 @@ const from16BitToFloat32 = (incomingData: Int16Array) => {
     return outputData;
 };
 
-/** Создает потоковый подгружаемый трек */
+/** Возвращает потоковый подгружаемый трек, который умеет себя проигрывать */
 export const createTrackStream = (
     ctx: AudioContext,
     {
@@ -90,12 +29,14 @@ export const createTrackStream = (
     }: {
         sampleRate?: number;
         numberOfChannels?: number;
+        // минимальная длина кусочка для воспроизведения, сек
         delay?: number;
         onPlay?: () => void;
         onEnd?: () => void;
         trackStatus?: 'stop' | 'play' | 'end';
     },
 ) => {
+    // очередь загруженных чанков (кусочков) трека
     const queue = createChunkQueue<AudioBufferSourceNode>();
     let extraByte: number | null = null;
     let status: 'stop' | 'play' | 'end' = trackStatus || 'stop';
@@ -131,6 +72,7 @@ export const createTrackStream = (
             return;
         }
 
+        // воспроизводим трек, если он полностью загрузился или длина загруженного больше задержки
         if (queue.loaded || queue.duration >= delay) {
             startTime = queue.length === 0 ? ctx.currentTime : startTime;
             const chunks = queue.popAll();
@@ -184,6 +126,7 @@ export const createTrackStream = (
         return source;
     };
 
+    /** добавляет чанк в очередь на воспроизведение */
     const write = (data: Uint8Array) => {
         // 44 байта - заголовок трека
         const slicePoint = firstChunk ? 44 : 0;
