+++ src/assistantSdk/voice/recognizers/asr/index.js
@@ -106,17 +106,31 @@
         Variables.decode = function decode(reader, length) {
             if (!(reader instanceof $Reader))
                 reader = $Reader.create(reader);
-            var end = length === undefined ? reader.len : reader.pos + length, message = new $root.Variables(), key;
+            var end = length === undefined ? reader.len : reader.pos + length, message = new $root.Variables(), key, value;
             while (reader.pos < end) {
                 var tag = reader.uint32();
                 switch (tag >>> 3) {
                 case 1:
-                    reader.skip().pos++;
                     if (message.variables === $util.emptyObject)
                         message.variables = {};
-                    key = reader.string();
-                    reader.pos++;
-                    message.variables[key] = reader.string();
+                    var end2 = reader.uint32() + reader.pos;
+                    key = "";
+                    value = "";
+                    while (reader.pos < end2) {
+                        var tag2 = reader.uint32();
+                        switch (tag2 >>> 3) {
+                        case 1:
+                            key = reader.string();
+                            break;
+                        case 2:
+                            value = reader.string();
+                            break;
+                        default:
+                            reader.skipType(tag2 & 7);
+                            break;
+                        }
+                    }
+                    message.variables[key] = value;
                     break;
                 default:
                     reader.skipType(tag & 7);
