+++ src/assistantSdk/voice/recognizers/asr/index.proto
